Updates Limb.py for Joint Trajectory Mode
Updates the Limb interface to utilize the
new TRAJECTORY_MODE in intera_core_msgs. This
new mode is a reworked RAW_POSITION_MODE but
requires Position, Velocity, and Acceleration
commands to execute a Trajectory.

The other modes (Position, Velocity, Torque)
now require commands to be populated in their
respective fields within the JointCommand msg.

Additionally, slight optimization in the JTAS
to execute commands strictly in the mode commanded.

Depends on
https://github.com/RethinkRobotics/intera_common/pull/7