Added Limb None queue_size and optimized traj
Adds optimizations to make control loop executions
more reliable in the amount of time each cycle
takes. This was done by forcing Synchronous publishing,
and controlling the execution of mode commands. Only
the lines of code required to execute a command per
command mode will execute in a control mode.