Checks initial point for jumps in JTAS
We need to check that the initial joint positions requested in the
trajectory requested is within the typical joint path thresholds,
in order to preven a faster-than-expected jump in the first JTAS
move.

A recent commit in MoveIt prevents this as well, setting a threshold
for aborting execution due to current joint angles.

Additionally, this commit tightens the default joint path error to
0.1 rad in the config