Added intera_interface Lights and lights_blink ex
This commit adds the intera_interface.Lights class
for setting and getting the internal state of each
light available on an intera robot. These lights
are individually controlable by color channel, but
are only "boolean" in nature. You can set a Blue
light OFF or ON, but not in between. All lights
and light colors are individually controlable.

- Adds intera_interface/src/intera_interface/lights.py
- Adds intera_examples/scripts/lights_blink.py

Reworked the settings import for some `intera_interface`
classes to be compatible with python 2.7.12