Updated Navigator Interface
Updates the navigator interface to utilize the IO framework.
This also gives more options for understanding the kind of
clicks received for the navigator (click, double click, long press)

Added the ability to register callbacks with this interface for
use when a navigator item value changes.

Additionally, this updates the joint_position_waypoints example
to give an idea how to make the robot interactive by using the
navigator. For an example that is purely navigator functionality,
see the updated intera_examples/navigator_io.py