Remove unnecessary Disable after Examples
In many Baxter examples, it made sense to disable the robot after
each example, if the robot was disabled to begin with. On Sawyer,
this workflow makes less sense, as the robot is nearly always enabled
unless there is an Error or E-Stop. For this reason, the Disabling
is removed from the examples.