Fix rospy_message_converter dep
This commit adds the rospy_message_converter dependency to
the package.xml and CMAKELISTS for the intera_motion_interface

It also fixes a small bug in ensure_path_to_file_exists,
which would throw an error if is was passed a filename without
a directory. Fixed with a simple check to handle empty strings.