Added script to stop the current motion trajectory
This commit adds a script that stops an active motion trajectory.
This is useful when the motion controller is running a long
trajectory that needs to be canceled.

I also made small modifications to the `MotionTrajectory` and
`MotionControllerActionClient` classes to make the `stop()` methods
behave more like the `send()` methods.