Provides default joint angles & nullspace bias to Cartesian script (#141)
This commit is a fix providing current joint configuration as nullspace joint bias.

Background: A recent change to the motion controller requires the nullspace joint bias to be provided with each waypoint. The current version of go_to_cartesian_pose.py does not provide the nullspace joint bias unless the joint angles are provided as an argument.
$ rosrun intera_examples go_to_cartesian_pose.py -p 0.4 -0.3 0.18 -o 0.0 1.0 0.0 0.0 -t right_hand
Should replicate the bug, resulting in
[ERROR] [1528747426.219658]: Motion controller failed to complete the trajectory with error CARTESIAN_INTERPOLATION_FAILED
being published to the terminal.