Add gitlab CI to validate sdk package
The CI add jobs for several ROS distributions

- Indigo
- Kinetic
- Melodic
- Noetic

It was necessary to handle the dependencies differently for each of the
ROS distributions.
For melodic and noetic, the joystick-drivers package is built from source,
while for indigo and kinetic it is obtained from apt-get.

In indigo the gitlab CI_JOB_TOKEN does not work, therefore the
intera_common is cloned from github instead.

Additional changes:
- Update copyright
- Bump package version to match intera 5.5.2

Co-authored-by: Adrian Zwiener <a.zwiener@rethinkrobotics.com>