Interface for Soft Touch Skill
The implementation is a wrapper around a simple action client, which is capable to communicate with the action server for the Soft Touch Skill. Some default parameters are already set, so that the user does not have to care about that. The interface can be embedded in the users python code.
---
GitLab: Intera-SDK #4